% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algaebase_query_v3.r
\name{algae_search}
\alias{algae_search}
\title{Compare a genus and species name against the algaebase online database}
\usage{
algae_search(genus, species = "", long = F)
}
\arguments{
\item{genus}{Character string}

\item{species}{Character string}

\item{long}{if TRUE, returns higher taxonomy (Kingdom through Family)}
}
\value{
A data.frame with the original name, binary flags indicating an exact match and an 
currently accepted name, potential accepted synonyms, and the best accepted match as a single 
string and as separate genus and species strings. Higher taxonomy may also be appended if long=T
}
\description{
Compare a genus and species name against the algaebase online database
}
\examples{
algae_search(genus='Anabaena',species='flos-aquae',long=TRUE)

}
\seealso{
\url{http://www.algaebase.org} for phytoplankton taxonomy database,
    \url{https://powellcenter.usgs.gov/geisha} for project information
}
