% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algaebase_output_parse.r
\name{algaebase_output_parse}
\alias{algaebase_output_parse}
\title{Helper function for parsing output from algaebase}
\usage{
algaebase_output_parse(x, field.name)
}
\arguments{
\item{x}{list object containing output from an algaebase query}

\item{field.name}{character string}
}
\value{
selected output variable as character vector
}
\description{
Helper function for parsing output from algaebase
}
