% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algaebase_query_v3.r
\name{bestmatch}
\alias{bestmatch}
\title{fuzzy partial matching between a scientific name and a list of possible matches}
\usage{
bestmatch(enteredName, possibleNames, maxErr = 3)
}
\arguments{
\item{enteredName}{Character string with name to check}

\item{possibleNames}{Character vector of possible matches}

\item{maxErr}{maximum number of different bits allowed for a partial match}
}
\value{
a character string with the best match, or 'multiplePartialMatches' if no best match
}
\description{
fuzzy partial matching between a scientific name and a list of possible matches
}
\examples{
possibleMatches=c('Viburnum edule','Viburnum acerifolia')
bestmatch(enteredName='Viburnum edulus',possibleNames=possibleMatches)
}
