% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_species_extract.r
\name{genus_species_extract}
\alias{genus_species_extract}
\title{Split a dataframe column with binomial name into genus and species columns.}
\usage{
genus_species_extract(phyto.df, phyto.name)
}
\arguments{
\item{phyto.df}{Name of data.frame object}

\item{phyto.name}{Character string: field in phyto.df containing species name.}
}
\value{
A data.frame with new character fields 'genus' and 'species'
}
\description{
Split a dataframe column with binomial name into genus and species columns.
}
\examples{
data(lakegeneva)
#example dataset with 50 rows

head(lakegeneva) #need to split the phyto_name column
new.lakegeneva=genus_species_extract(lakegeneva,'phyto_name')

head(new.lakegeneva)
}
