% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apikey_from_Renviron.r
\name{get_apikey}
\alias{get_apikey}
\title{Get value of algaebase API key from Environment variable
Return an error if variable not set.}
\usage{
get_apikey()
}
\value{
api key as character string (invisibly)
}
\description{
Get value of algaebase API key from Environment variable
Return an error if variable not set.
}
