% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakegeneva.R
\docType{data}
\name{lakegeneva}
\alias{lakegeneva}
\title{example dataset from lake Geneva, Switzerland (50 lines)}
\format{A data frame with columns:
\describe{
 \item{lake}{lake name}
 \item{date_dd_mm_yy}{date dd-mm-yy format}
 \item{phyto_name}{phytoplankton species binomial name}
}}
\usage{
data(lakegeneva)
}
\description{
example dataset from lake Geneva, Switzerland (50 lines)
}
\keyword{datasets}
