% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_mat.R
\name{mean_naomit}
\alias{mean_naomit}
\title{Compute mean value while ignoring NA's}
\usage{
mean_naomit(x)
}
\arguments{
\item{x}{A numeric vector that may contain NA's}
}
\value{
the mean value
}
\description{
Compute mean value while ignoring NA's
}
\examples{
data(lakegeneva)
#example dataset with 50 rows

mean_naomit(lakegeneva$biovol_um3_ml)
}
