% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfg.csr.r
\docType{data}
\name{mfg.csr}
\alias{mfg.csr}
\title{MFG-CSR correspondence based on CSR-trait relationships in Reynolds et al. 1988
and MFG-trait relationships in Salmaso et al. 2015}
\format{A data frame with columns:
\describe{
 \item{MFG.number}{shortened MFG designation}
 \item{MFG}{full MFG name from Salmaso et al. 2015}
 \item{CSR}{CSR classification including intermediate classes}
}}
\usage{
data(mfg.csr)
}
\description{
MFG-CSR correspondence based on CSR-trait relationships in Reynolds et al. 1988
and MFG-trait relationships in Salmaso et al. 2015
}
\keyword{datasets}
