% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfg_csr_library.r
\docType{data}
\name{mfg_csr_library}
\alias{mfg_csr_library}
\title{MFG-CSR correspondence based on CSR-trait relationships in Reynolds et al. 1988
and MFG-trait relationships in Salmaso et al. 2015}
\format{
A data frame with columns:
\describe{
 \item{MFG}{full MFG name from Salmaso et al. 2015}
 \item{CSR}{CSR classification including intermediate classes}
}
}
\usage{
data(mfg_csr_library)
}
\description{
MFG-CSR correspondence based on CSR-trait relationships in Reynolds et al. 1988
and MFG-trait relationships in Salmaso et al. 2015
}
\keyword{datasets}
