% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto_convert_df.r
\name{phyto_convert_df}
\alias{phyto_convert_df}
\title{Wrapper function to apply species_phyto_convert() across a data.frame}
\usage{
phyto_convert_df(phyto.df, flag = 1)
}
\arguments{
\item{phyto.df}{Name of data.frame. Must have character fields named 'genus' and 'species'}

\item{flag}{Resolve ambiguous mfg: 1 = return(NA),2 = manual selection}
}
\value{
a single MFG classification as character string
}
\description{
Wrapper function to apply species_phyto_convert() across a data.frame
}
\examples{
data(lakegeneva)
#example dataset with 50 rows

new.lakegeneva=genus_species_extract(lakegeneva,'phyto_name')
new.lakegeneva=phyto_convert_df(new.lakegeneva)
head(new.lakegeneva)
}
