% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_algaebase_apikey_header.r
\name{set_algaebase_apikey_header}
\alias{set_algaebase_apikey_header}
\title{Add algaebase API key to curl handle}
\usage{
set_algaebase_apikey_header(apikey = NULL)
}
\arguments{
\item{apikey}{character string with valid key}
}
\value{
curl handle object
}
\description{
Add algaebase API key to curl handle
}
