% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mfg.library.r
\docType{data}
\name{species.mfg.library}
\alias{species.mfg.library}
\title{Trait-based MFG classifications for common Eurasion/North American phytoplankton species.
See accompanying manuscript for sources}
\format{A data frame with columns:
\describe{
 \item{genus}{genus name}
 \item{species}{species name}
 \item{MFG}{corresponding MFG classification based on Salmaso et al. 2015}
 }}
\usage{
data(species.mfg.library)
}
\description{
Trait-based MFG classifications for common Eurasion/North American phytoplankton species.
See accompanying manuscript for sources
}
\keyword{datasets}
