% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mfg_library.r
\docType{data}
\name{species_mfg_library}
\alias{species_mfg_library}
\title{Trait-based MFG classifications for common Eurasion/North American phytoplankton species.
See accompanying manuscript for sources}
\format{
A data frame with columns:
\describe{
 \item{genus}{genus name}
 \item{species}{species name}
 \item{MFG}{corresponding MFG classification based on Salmaso et al. 2015}
 \item{source}{literature or online source for MFG classification}
 }
}
\usage{
data(species_mfg_library)
}
\description{
Trait-based MFG classifications for common Eurasion/North American phytoplankton species.
See accompanying manuscript for sources
}
\references{
{Algaebase} \url{https://www.algaebase.org}

{Phycokey} \url{https://www.cfb.unh.edu/phycokey/phycokey.htm}

{Western Diatoms of North America} \url{https://diatoms.org}

{CyanoDB 2} \url{http://www.cyanodb.cz/}

{Nordic Microalgae} \url{https://nordicmicroalgae.org}

{Phytopedia} \url{https://phytoplankton.eoas.ubc.ca/}

{Kapustin, D., Sterlyagova, I. and Patova, E., 2019. Morphology of Chrysastrella paradoxa stomatocysts from the Subpolar Urals (Russia) with comments on related morphotypes. Phytotaxa, 402(6), pp.295-300.}
}
\keyword{datasets}
