% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_phyto_convert.r
\name{species_phyto_convert}
\alias{species_phyto_convert}
\title{Conversion of a single genus and species name to a single MFG}
\usage{
species_phyto_convert(genus, species, flag = 1)
}
\arguments{
\item{genus}{Character string: genus name}

\item{species}{Character string: species name}

\item{flag}{Resolve ambiguous mfg: 1 = return(NA),2= manual selection}
}
\value{
a single MFG classification as character string
}
\description{
Conversion of a single genus and species name to a single MFG
}
\examples{
species_phyto_convert('Scenedesmus','bijuga')
#returns "11a-NakeChlor"
}
