% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_to_mfg.r
\name{species_to_mfg}
\alias{species_to_mfg}
\title{Conversion of a single genus and species name to a single MFG. Uses species.mfg.library}
\usage{
species_to_mfg(genus, species, flag = 1)
}
\arguments{
\item{genus}{Character string: genus name}

\item{species}{Character string: species name}

\item{flag}{Resolve ambiguous mfg: 1 = return(NA),2= manual selection}
}
\value{
a single MFG classification as character string
}
\description{
Conversion of a single genus and species name to a single MFG. Uses species.mfg.library
}
\examples{
species_to_mfg('Scenedesmus','bijuga')
#returns "11a-NakeChlor"
}
