% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_list_algaebase.r
\name{spp_list_algaebase}
\alias{spp_list_algaebase}
\title{Wrapper for applying algae_search function to a data.frame that contains phytoplankton species}
\usage{
spp_list_algaebase(phyto.df, phyto.name = 1, lakename = "",
  long = FALSE, write = FALSE)
}
\arguments{
\item{phyto.df}{data.frame containing a character column with binomial names}

\item{phyto.name}{Name or number of column that contains binomial names}

\item{lakename}{Character string for naming output files}

\item{long}{TRUE/FALSE: should higher taxonomy (Kingdom:Family) be included in output?}

\item{write}{TRUE/FALSE: should output be written as .csv file?}
}
\value{
A character string of the species' morphofunctional group
}
\description{
Wrapper for applying algae_search function to a data.frame that contains phytoplankton species
}
\examples{
data(lakegeneva)
lakegeneva=lakegeneva[1:3,] ##use 3 rows for testing
lakegeneva.algaebase<-
spp_list_algaebase(lakegeneva,phyto.name='phyto_name',long=TRUE,write=FALSE)

}
\seealso{
\url{http://www.algaebase.org} for up-to-date phytoplankton taxonomy,
    \url{https://powellcenter.usgs.gov/geisha} for project information
}
