% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitranges.r
\docType{data}
\name{traitranges}
\alias{traitranges}
\title{surface/volume ratio and max linear dimension criteria for CSR
From Reynolds 1988 and Reynolds 2006}
\format{A data frame with columns:
\describe{
 \item{Measurement}{measurement type}
 \item{C.min}{minimum value for C}
 \item{S.min}{minimum value for S}
 \item{R.min}{minimum value for R}
 \item{C.max}{maximum value for C}
 \item{S.max}{maximum value for S}
 \item{R.max}{maximum value for R}
 \item{units}{units of measurement}
 \item{source}{source for criteria}
}}
\usage{
data(traitranges)
}
\description{
surface/volume ratio and max linear dimension criteria for CSR
From Reynolds 1988 and Reynolds 2006
}
\keyword{datasets}
