% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_to_csr.r
\name{traits_to_csr}
\alias{traits_to_csr}
\title{Assign phytoplankton species to CSR functional groups, based on surface to volume ratio and
maximum linear dimension ranges proposed by Reynolds et al. 1988;2006}
\usage{
traits_to_csr(sav, msv)
}
\arguments{
\item{sav}{numeric estimate of cell or colony surface/volume ratio in um^2/um^3}

\item{msv}{numeric product of surface/volume ratio multiplied by maximum linear dimension (msv; um)}
}
\value{
a character string with one of 3 return values: C,S, or R
}
\description{
Assign phytoplankton species to CSR functional groups, based on surface to volume ratio and
maximum linear dimension ranges proposed by Reynolds et al. 1988;2006
}
\examples{

traits_to_csr(sav=0.2,msv=10)


}
\seealso{
/url{https://powellcenter.usgs.gov/geisha} for project information
}
