% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_search_itis.r
\name{genus_search_itis}
\alias{genus_search_itis}
\title{Wrapper function for several functions in ritis::
Searches ITIS database for matches to a genus name}
\usage{
genus_search_itis(genus, higher = FALSE)
}
\arguments{
\item{genus}{Character string. genus name to search for in ITIS}

\item{higher}{Boolean. If TRUE, add higher taxonomic classifications to output}
}
\value{
input data.frame with matches, current accepted names, synonyms, and higher taxonomy
}
\description{
Wrapper function for several functions in ritis::
Searches ITIS database for matches to a genus name
}
\examples{
genus='Anabaena'
genus_search_itis(genus,higher=FALSE)
}
