% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakegeneva.R
\docType{data}
\name{lakegeneva}
\alias{lakegeneva}
\title{example dataset from lake Geneva, Switzerland}
\format{
A data frame with columns:
\describe{
 \item{lake}{lake name}
 \item{phyto_name}{phytoplankton species name}
 \item{month}{month of sampling}
 \item{year}{year of sampling}
 \item{date_dd_mm_yy}{date of sampling}
 \item{biovol_um3_ml}{biovolume}
}
}
\usage{
data(lakegeneva)
}
\description{
example dataset from lake Geneva, Switzerland
}
\keyword{datasets}
